move archive utilities to localstack.utils.archives
