+++ localstack/utils/common.py
@@ -14,12 +14,10 @@ import platform
 import re
 import subprocess
 import sys
-import tarfile
 import tempfile
 import threading
 import time
 import uuid
-import zipfile
 from datetime import date, datetime, timezone, tzinfo
 from json import JSONDecodeError
 from multiprocessing.dummy import Pool
@@ -35,6 +33,9 @@ import localstack.utils.run
 from localstack import config
 from localstack.config import DEFAULT_ENCODING
 from localstack.constants import ENV_DEV
+
+# TODO: remove imports from here (need to update any client code that imports these from utils.common)
+from localstack.utils.archives import is_zip_file, untar, unzip  # noqa
 
 # TODO: remove imports from here (need to update any client code that imports these from utils.common)
 from localstack.utils.collections import (  # noqa
@@ -75,6 +76,8 @@ from localstack.utils.files import (  # noqa
     rm_rf,
     save_file,
 )
+
+# TODO: remove imports from here (need to update any client code that imports these from utils.common)
 from localstack.utils.generic.number_utils import format_number, is_number
 
 # TODO: remove imports from here (need to update any client code that imports these from utils.common)
@@ -1235,57 +1238,6 @@ def kill_process_tree(parent_pid):
         except Exception:
             pass
     parent.kill()
-
-
-def is_zip_file(content):
-    stream = io.BytesIO(content)
-    return zipfile.is_zipfile(stream)
-
-
-def unzip(path, target_dir, overwrite=True):
-    is_in_debian = is_debian()
-    if is_in_debian:
-        # Running the native command can be an order of magnitude faster in Alpine on Travis-CI
-        flags = "-o" if overwrite else ""
-        flags += " -q"
-        try:
-            return run("cd %s; unzip %s %s" % (target_dir, flags, path), print_error=False)
-        except Exception as e:
-            error_str = truncate(str(e), max_length=200)
-            LOG.info(
-                'Unable to use native "unzip" command (using fallback mechanism): %s', error_str
-            )
-
-    try:
-        zip_ref = zipfile.ZipFile(path, "r")
-    except Exception as e:
-        LOG.warning("Unable to open zip file: %s: %s", path, e)
-        raise e
-
-    def _unzip_file_entry(zip_ref, file_entry, target_dir):
-        """Extracts a Zipfile entry and preserves permissions"""
-        out_path = os.path.join(target_dir, file_entry.filename)
-        if is_in_debian and os.path.exists(out_path) and os.path.getsize(out_path) > 0:
-            # this can happen under certain circumstances if the native "unzip" command
-            # fails with a non-zero exit code, yet manages to extract parts of the zip file
-            return
-        zip_ref.extract(file_entry.filename, path=target_dir)
-        perm = file_entry.external_attr >> 16
-        # Make sure to preserve file permissions in the zip file
-        # https://www.burgundywall.com/post/preserving-file-perms-with-python-zipfile-module
-        os.chmod(out_path, perm or 0o777)
-
-    try:
-        for file_entry in zip_ref.infolist():
-            _unzip_file_entry(zip_ref, file_entry, target_dir)
-    finally:
-        zip_ref.close()
-
-
-def untar(path, target_dir):
-    mode = "r:gz" if path.endswith("gz") else "r"
-    with tarfile.open(path, mode) as tar:
-        tar.extractall(path=target_dir)
 
 
 def is_root():
